requirements.txt
instagrapi==2.0.0
pyautogen==0.2.31
openai==1.35.7
python-dotenv==1.0.1
flask==3.1.2
httpx==0.24.1